XP-8 -- added promotional image at end of pdf

+++ xpapers/tasks.py
@@ -10,6 +10,7 @@ from django.core.files import File
 from django.core.files.storage import default_storage
 from django.core.files.base import ContentFile
 from django.conf import settings
+from PIL import Image, ImageDraw, ImageFont
 from reportlab.pdfgen import canvas
 from reportlab.lib.pagesizes import letter
 from dashboard.models import Post, PostFiles
@@ -78,9 +79,39 @@ def celery_pdf_watermark(*args, **kwargs):
 @shared_task
 def celery_images_to_pdf(*args, **kwargs):
     print('####~~celery-images_to_pdf~~####')
+    user_login = False
+    static_root = os.path.join(settings.BASE_DIR, 'static')
+    promotion_image_path = os.path.join(static_root,
+                                        'utils/upload_papers_utils/pdf_promotion_image.jpg')
+    coval_font_path = os.path.join(static_root,
+                                   'utils/upload_papers_utils/Coval-Black.otf')
     image_list = args[0]['image_list']
     user = args[0]['user']
     post_id = args[0]['post_id']
+    if user:
+        # Adding name of user in pdf promotion image
+        try:
+            image = Image.open(promotion_image_path)
+            draw = ImageDraw.Draw(image)
+            font = ImageFont.truetype(coval_font_path, size=45)
+            (x, y) = (480, 400)
+            name = "@" + user
+            color = 'rgb(0, 0, 0)' # white color
+            draw.text((x, y), name, fill=color, font=font)
+            temp_promotion_file = NamedTemporaryFile(delete=False, suffix='.jpg')
+            image.save(temp_promotion_file)
+            image_list.append(temp_promotion_file.name)
+            user_login = True
+        except Exception as e:
+            print('$$~~ ERROR ~ PDF~PROMOTION~IMAGE ~ ERROR ~~$$ POST-ID -%s- Exception %s' % (post_id, e))
+            with push_scope() as scope:
+                scope.set_tag("Place", "celery-PDF-Promotion-Image")
+                scope.set_extra("Post ID", post_id)
+                scope.set_extra('User', user)
+                scope.level = 'error'
+                capture_exception(e)
+    else:
+        image_list.append(promotion_image_path)
     try:
         temp_file = NamedTemporaryFile(delete=False, suffix='.pdf')
         with open(temp_file.name, "wb") as f:
@@ -92,6 +123,10 @@ def celery_images_to_pdf(*args, **kwargs):
             }
         celery_pdf_watermark(celery_data)
         for img in image_list:
+            if not user_login and img == image_list[-1]:
+                # Not to delete static pdf promotion image
+                # if user is not loged in
+                continue
             os.remove(img)
     except Exception as e:
         print('$$~~ ERROR ~ IMAGE~TO~PDF ~ ERROR ~~$$ POST-ID -%s- Exception %s' % (post_id, e))
